requirements.txt
pydantic>=2.0
pydantic-ai>=0.1.8
openai>=1.0
requests
beautifulsoup4
readability-lxml
